[nxp][examples][k32w0] Update build system after moving some components
